add option to migrate

+++ inc/class/WPCasamaMigration.php
@@ -6,55 +6,60 @@
 		public $user_id;
 		
 		function __construct() {
-			//add_action('admin_init', array($this, 'wpcasama_select_old_alert'));
+			add_action('admin_init', array($this, 'wpcasama_select_old_alert'));
 		}
 		
 		function wpcasama_select_old_alert(){
-			global $wpdb;
+			$migrate = wpsight_get_option('wpcasama_migration');
 			
-			$table_name = $wpdb->prefix . 'wpcasama_mailalert';
-			$alerts = $wpdb->get_results("SELECT * FROM $table_name");
-			
-			
-			
-			foreach ( $alerts as $alert ){
+			if ( !defined( 'WPCASAMA_MIGRATED') && $migrate = '1' ) {
+				global $wpdb;
 				
-				//var_dump($alert);
-				$userdata = array(
-					'user_login' => $alert->name,
-					'user_email' => $alert->email,
-					'user_nicename' => $alert->name,
-				
-				);
-				
-				$user_exists = get_user_by('email', $userdata['user_email']);
+				$table_name = $wpdb->prefix . 'wpcasama_mailalert';
+				$alerts     = $wpdb->get_results( "SELECT * FROM $table_name" );
 				
 				
-				if ( $user_exists == false ){
-					$this->user_id = wp_insert_user( $userdata );
-				} else {
-					$this->user_id = $user_exists->ID;
+				foreach ( $alerts as $alert ) {
+					
+					//var_dump($alert);
+					$userdata = array(
+						'user_login'    => $alert->name,
+						'user_email'    => $alert->email,
+						'user_nicename' => $alert->name,
+					
+					);
+					
+					$user_exists = get_user_by( 'email', $userdata['user_email'] );
+					
+					
+					if ( $user_exists == false ) {
+						$this->user_id = wp_insert_user( $userdata );
+					} else {
+						$this->user_id = $user_exists->ID;
+					}
+					
+					$postarr = array(
+						'post_author' => $this->user_id,
+						'post_title'  => $alert->name . '-' . $alert->city,
+						'post_status' => 'publish',
+						'post_type'   => 'wpcasa-mail-alerte',
+						'post_date'   => $alert->subscription,
+						'meta_input'  => array(
+							'wpcasama_phone'     => $alert->tel,
+							'wpcasama_city'      => $alert->city,
+							'wpcasama_min_price' => intval( $alert->min_price ),
+							'wpcasama_max_price' => intval( $alert->max_price ),
+						),
+					
+					);
+					
+					$alert_created = wp_insert_post( $postarr );
+					
+					$wpdb->query( "DROP TABLE IF EXISTS $table_name" );
+					
+					define('WPCASAMA_MIGRATED', true);
+					
 				}
-				
-				$postarr = array(
-					'post_author'   => $this->user_id,
-					'post_title'    => $alert->name . '-' . $alert->city,
-					'post_status'   => 'publish',
-					'post_type' =>  'wpcasa-mail-alerte',
-					'post_date' =>  $alert->subscription,
-					'meta_input' => array(
-						'wpcasama_phone' => $alert->tel,
-						'wpcasama_city' => $alert->city,
-						'wpcasama_min_price' => intval($alert->min_price),
-						'wpcasama_max_price' => intval($alert->max_price),
-					),
-				
-				);
-				
-				$alert_created = wp_insert_post($postarr);
-				
-				$wpdb->query( "DROP TABLE IF EXISTS $table_name" );
-				
 			}
